apps/agents/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-adk==0.1.0
google-generativeai>=0.7.0
pydantic>=2.7.0
pydantic-settings>=2.5.2
httpx>=0.28.1
python-dotenv>=1.0.1
